client command line

+++ logStorage/logStorage.go
@@ -7,7 +7,7 @@ import (
 type LogStorage interface {
 	Create(topic string) (bool, error)
 	Drop(topic string) (bool, error)
-	Write(topicMessage *TopicMessage) (int, error)
+	Write(topicMessage *TopicMessage) (uint64, error)
 	WriteBatch(topicBatchMessage *TopicBatchMessage) (int, error)
 	ReadFromBeginning(logChan chan LogEntry, wg *sync.WaitGroup, topic string) error
 	ReadBatchFromBeginning(logChan chan LogEntryBatch, wg *sync.WaitGroup, topic string, batchSize int) error
